+++ tools/marisa-build.cc
@@ -202,5 +202,6 @@ int main(int argc, char *argv[]) {
       }
     }
   }
-  return build(cmdopt.argv + cmdopt.optind, cmdopt.argc - cmdopt.optind);
+  return build(cmdopt.argv + cmdopt.optind,
+      static_cast<std::size_t>(cmdopt.argc - cmdopt.optind));
 }
